+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 12.7
 - SQL allow array in $where_columns: WHERE COLUMN IN(val1,val2) in DBUpsert.php
 - Fix #363 Pull-Down field: add "No Value" option in AssignOtherInfo.php
 - Fix #363 Checkbox Field: save unchecked state in AssignOtherInfo.php
+- Fix #363 Auto Pull-Down field: add "-Edit-" option in AssignOtherInfo.php
 
 Changes in 12.6
 ---------------